# Conflicts:
#	src/main/java/gregtech/api/metatileentity/implementations/MTEMultiBlockBase.java
#	src/main/java/gregtech/api/recipe/RecipeMaps.java
#	src/main/java/gregtech/api/util/GTRecipeConstants.java
#	src/main/java/gregtech/common/blocks/BlockCasings12.java
#	src/main/java/gregtech/common/blocks/ItemCasings12.java
#	src/main/java/gregtech/loaders/preload/LoaderMetaTileEntities.java
#	src/main/resources/assets/gregtech/lang/en_US.lang

+++ src/main/java/gtPlusPlus/core/container/ContainerVolumetricFlaskSetter.java
@@ -160,7 +160,7 @@ public class ContainerVolumetricFlaskSetter extends Container {
     public void updateProgressBar(int par1, int par2) {
         super.updateProgressBar(par1, par2);
         if (par1 == 0) {
-            mCustomValue = (short) par2;
+            mCustomValue = par2;
         }
     }
